% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aveTies.r
\name{aveTies}
\alias{aveTies}
\title{Internal function: average ties}
\usage{
aveTies(RANKS, normExprs)
}
\arguments{
\item{RANKS}{matrix of ranks}

\item{normExprs}{normalized values}
}
\description{
Internal function: average ties
}

