% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispf.r
\name{dispf}
\alias{dispf}
\title{Generate genewise dispersion}
\usage{
dispf(mu, a, chiDF)
}
\arguments{
\item{mu}{base line counts}

\item{a}{asymptotic dispersion}

\item{chiDF}{degree of freedom gene specific disp. factor}
}
\description{
Generate genewise dispersion
}

