% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterCounts.r
\name{filterCounts}
\alias{filterCounts}
\title{Filter out low counts}
\usage{
filterCounts(counts, thresh = 1, minSamples = 2)
}
\arguments{
\item{counts}{counts matrix (do not add 1)}

\item{thresh}{minimum counts per million to determine expression (default=1)}

\item{minSamples}{minimum number of samples where gene is required to be expressed.
This should be set to the numer of samples in the smallest group of interest. (default=2)}
}
\description{
Filter out low counts
}

