% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitCoeffs.r
\name{fitCoeffs}
\alias{fitCoeffs}
\title{Fit alpha group coefficients
Notes:
1. Currently using ridge regression
Output:
returns a list: (1.) a matrix of fitted group means for each gene (2.) design matrix}
\usage{
fitCoeffs(alpha, groups, lambda, levels = NULL)
}
\arguments{
\item{alpha}{matrix of reference adjusted quantiles (genes by samples)}

\item{groups}{groups factor / character}

\item{lambda}{shrinkage parameter}

\item{levels}{(optional) specify order of levels}
}
\description{
Fit alpha group coefficients
Notes:
1. Currently using ridge regression
Output:
returns a list: (1.) a matrix of fitted group means for each gene (2.) design matrix
}

