% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fStat.r
\name{fStat}
\alias{fStat}
\title{Perform row f-tests
Notes:
Output:
A list: (1.) fstat (2.) df1 (3.) df2}
\usage{
fStat(alpha, groups, levels = NULL)
}
\arguments{
\item{alpha}{matrix of median adjusted quantiles (genes by samples)}

\item{groups}{groups factor / character}

\item{levels}{(optional) specify order of levels}
}
\description{
Perform row f-tests
Notes:
Output:
A list: (1.) fstat (2.) df1 (3.) df2
}

