% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/medScal.r
\name{medScal}
\alias{medScal}
\title{Median scale data
Notes:
1.
Output:
Median scaled data}
\usage{
medScal(data)
}
\arguments{
\item{data}{(positive values: raw intensities or pseudo-counts)}
}
\description{
Median scale data
Notes:
1.
Output:
Median scaled data
}

