% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predicted.r
\name{predicted}
\alias{predicted}
\title{Compute predicted alpha values
Notes:
1. Takes in results of fitCoeffs()
# Output:
matrix of predicted values}
\usage{
predicted(design, betas)
}
\arguments{
\item{design}{= design matrix}

\item{betas}{= matrix of betas}
}
\description{
Compute predicted alpha values
Notes:
1. Takes in results of fitCoeffs()
# Output:
matrix of predicted values
}

