% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qshrink.r
\name{qshrink}
\alias{qshrink}
\title{Quantile shrinkage normalization}
\usage{
qshrink(exprs, groups, refType = "mean", groupLoc = "mean", window = 99,
  verbose = FALSE, groupCol = NULL, plot = FALSE, ...)
}
\arguments{
\item{exprs}{for counts use log2(raw counts + 1)), for MA use log2(raw intensities)}

\item{groups}{groups to which samples belong (character vector)}

\item{refType}{type of quantile reference (defualt="mean")}

\item{groupLoc}{type of group location estimate (default="mean")}

\item{window}{window size for running median}

\item{verbose}{show info}

\item{groupCol}{character vector indicating the group color for each sample}

\item{plot}{plot weights? (default=FALSE)}

\item{...}{plot options}
}
\description{
Quantile shrinkage normalization
}

