% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{qsmooth-class}
\alias{qsmooth-class}
\title{the qsmooth class}
\value{
\code{qsmoothWeights} returns the qsmooth weights and 
\code{qsmoothData} returns the qsmooth normalized data
}
\description{
Objects of this class store all 
the values needed information to work with a
qsmooth object
}
\section{Slots}{

\describe{
\item{\code{qsmoothWeights}}{qsmooth weights}

\item{\code{qsmoothData}}{qsmooth normalized data}
}}

\examples{
library(SummarizedExperiment)
library(bodymapRat)
bm_dat <- bodymapRat()

# select lung and liver samples, stage 21 weeks, and bio reps
keep_columns = (colData(bm_dat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bm_dat)$stage == 21) & 
         (colData(bm_dat)$techRep == 1)
keep_rows = rowMeans(assay(bm_dat)) > 10 # Filter out low counts
bm_dat <- bm_dat[keep_rows,keep_columns]

qs_norm <- qsmooth(object = assay(bm_dat), 
                  group_factor = colData(bm_dat)$organ)

}
