% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/residuals.r
\name{residuals}
\alias{residuals}
\title{Compute residuals
Notes:
1.
Output:
returns a matrix of residuals}
\usage{
residuals(y, groups, lambda, levels = NULL)
}
\arguments{
\item{y}{matrix of median adjusted quantiles (genes by samples)}

\item{groups}{groups factor / character}

\item{lambda}{shrinkage parameter}

\item{levels}{(optional) specify order of levels}
}
\description{
Compute residuals
Notes:
1.
Output:
returns a matrix of residuals
}

