% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaleData.r
\name{scaleData}
\alias{scaleData}
\title{Scale normalization}
\usage{
scaleData(exprs, type = "trim.mean")
}
\arguments{
\item{exprs}{counts + 1 matrix or PM values}

\item{type}{type of scaling (default = "trim.mean")}
}
\description{
Scale normalization
}

