% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scals.r
\name{scals}
\alias{scals}
\title{Compute sample quantiles, the quantile reference, and alpha matrix.
Notes:
1.
Output:
A list: (1.) log2(sample quantiles) (2.) log2(quantile ref.) (3.) alphas}
\usage{
scals(data)
}
\arguments{
\item{data}{(positive values: raw intensities or pseudo-counts)}
}
\description{
Compute sample quantiles, the quantile reference, and alpha matrix.
Notes:
1.
Output:
A list: (1.) log2(sample quantiles) (2.) log2(quantile ref.) (3.) alphas
}

