% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{qsmooth-class}
\alias{qsmooth-class}
\title{the qsmooth class}
\value{
\code{qsmoothWeights} returns the qsmooth weights and 
\code{qsmoothData} returns the qsmooth normalized data
}
\description{
Objects of this class store all 
the values needed information to work with a
qsmooth object
}
\section{Slots}{

\describe{
\item{\code{qsmoothWeights}}{qsmooth weights}

\item{\code{qsmoothData}}{qsmooth normalized data}
}}

\examples{
dat <- cbind(matrix(rnorm(1000), nrow=100, ncol=10), 
             matrix(rnorm(1000, .1, .7), nrow=100, ncol=10))
dat_qs <- qsmooth(object = dat, 
                  group_factor = rep(c(0,1), each=10))

}
