% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qsmooth.r
\name{qsmooth}
\alias{qsmooth}
\title{Quantile shrinkage normalization}
\usage{
qsmooth(exprs, groups, norm.factors = NULL, plot = FALSE, window = 0.05)
}
\arguments{
\item{exprs}{for counts use log2(raw counts + 1)), for MA use log2(raw intensities)}

\item{groups}{groups to which samples belong (character vector)}

\item{norm.factors}{scaling normalization factors}

\item{plot}{plot weights? (default=FALSE)}

\item{window}{window size for running median as a fraction on the number of rows of exprs}
}
\description{
Quantile shrinkage normalization
}

