% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{qsmoothData}
\alias{qsmoothData}
\alias{qsmoothData,qsmooth-method}
\alias{qsmoothData.qsmooth}
\title{Accessors for the 'qsmoothData' slot of a qsmooth object.}
\usage{
qsmoothData(object)

\S4method{qsmoothData}{qsmooth}(object)

\S4method{qsmoothData}{qsmooth}(object)
}
\arguments{
\item{object}{a \code{qsmooth} object}

\item{...}{other}
}
\description{
Accessors for the 'qsmoothData' slot of a qsmooth object.
}

