% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{qsmoothData}
\alias{qsmoothData}
\alias{qsmoothData,qsmooth-method}
\title{Generic function that returns the qsmooth normalized data}
\usage{
qsmoothData(object)

\S4method{qsmoothData}{qsmooth}(object)
}
\arguments{
\item{object}{an object of class \code{qsmooth}.}
}
\value{
The normalized data after applying 
smoothed quantile normalization.
}
\description{
Given a qsmooth object, this function returns the 
qsmooth normalized data

Accessors for the 'qsmoothData' slot of a qsmooth object.
}
\examples{
library(SummarizedExperiment)
library(bodymapRat)
data(bodymapRat)

# select lung and liver samples, stage 21 weeks, and bio reps
keepColumns = (colData(bodymapRat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bodymapRat)$stage == 21) & 
         (colData(bodymapRat)$techRep == 1)
keepRows = rowMeans(assay(bodymapRat)) > 10 # Filter out low counts
bodymapRat <- bodymapRat[keepRows,keepColumns]

qsNorm <- qsmooth(object = assay(bodymapRat), 
                  groupFactor = colData(bodymapRat)$organ)
qsmoothData(qsNorm)

}
