% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qsmoothPlotWeights}
\alias{qsmoothPlotWeights}
\title{Plot weights from \code{qsmooth} function.}
\usage{
qsmoothPlotWeights(object, xLab = "quantiles", yLab = "weights",
  mainLab = "qsmooth weights")
}
\arguments{
\item{object}{a qsmooth object from \code{qsmooth}}

\item{xLab}{label for x-axis. Default is "quantiles"}

\item{yLab}{label for y-axis. Default is "weights"}

\item{mainLab}{title of plot. Default is "qsmooth weights"}
}
\value{
A scatterplot will be created showing the 
\code{qsmoothWeights} along the y-axis and the 
quantiles on the x-axis.
}
\description{
This function plots a scatterplot
showing the \code{qsmoothWeights} along the y-axis
and the quantiles on the x-axis.
}
\examples{
library(SummarizedExperiment)
library(bodymapRat)
bm_dat <- bodymapRat()

# select lung and liver samples, stage 21 weeks, and bio reps
keep_columns = (colData(bm_dat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bm_dat)$stage == 21) & 
         (colData(bm_dat)$techRep == 1)
keep_rows = rowMeans(assay(bm_dat)) > 10 # Filter out low counts
bm_dat <- bm_dat[keep_rows,keep_columns]

qs_norm <- qsmooth(object = assay(bm_dat), 
                  group_factor = colData(bm_dat)$organ)
qsmoothPlotWeights(qs_norm)
}
