% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qsmoothPlotWeights}
\alias{qsmoothPlotWeights}
\title{Plot weights from \code{qsmooth} function.}
\usage{
qsmoothPlotWeights(object, savePlot = FALSE, xLab = NULL, yLab = NULL,
  mainLab = NULL, binWidth = NULL)
}
\arguments{
\item{object}{a qsmooth object from \code{qsmooth}}

\item{savePlot}{a TRUE/FALSE object argument
determining if the plot will be saved for further use or
immediately displayed on the screen.}

\item{xLab}{label for x-axis}

\item{yLab}{label for y-axis}

\item{mainLab}{title of plot}

\item{binWidth}{binwidth or histogram. Default is the stat_bin default.}
}
\value{
A histogram will be plotted containing the null 
test statistics when using boostrapped samples. 
The red line is the observed test statistic
\code{quantroStat} from \code{qsmooth()}.
}
\description{
This function plots the histogram of the
null test statistics from permutation test in the \code{qsmooth}
function.
}
\examples{
\donttest{
library(minfi)
data(flowSorted)
p <- getBeta(flowSorted, offset = 100)
pd <- pData(flowSorted)

library(doParallel)
registerDoParallel(cores=4)
qtest <- quantro(p, pd$CellType, B = 100)
quantroPlot(qtest)
}
}

