% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qsmoothPlotWeights}
\alias{qsmoothPlotWeights}
\title{Plot weights from \code{qsmooth} function.}
\usage{
qsmoothPlotWeights(object, savePlot = FALSE, xLab = NULL,
  yLab = NULL, mainLab = NULL, binWidth = NULL)
}
\arguments{
\item{object}{a qsmooth object from \code{qsmooth}}

\item{savePlot}{a TRUE/FALSE object argument
determining if the plot will be saved for further use or
immediately displayed on the screen.}

\item{xLab}{label for x-axis}

\item{yLab}{label for y-axis}

\item{mainLab}{title of plot}

\item{binWidth}{binwidth or histogram. Default is the stat_bin default.}
}
\value{
A histogram will be plotted containing the null 
test statistics when using boostrapped samples. 
The red line is the observed test statistic
\code{quantroStat} from \code{qsmooth()}.
}
\description{
This function plots the histogram of the
null test statistics from permutation test in the \code{qsmooth}
function.
}
\examples{
library(SummarizedExperiment)
library(bodymapRat)
data(bodymapRat)

# select lung and liver samples, stage 21 weeks, and bio reps
keepColumns = (colData(bodymapRat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bodymapRat)$stage == 21) & 
         (colData(bodymapRat)$techRep == 1)
keepRows = rowMeans(assay(bodymapRat)) > 10 # Filter out low counts
bodymapRat <- bodymapRat[keepRows,keepColumns]

qsNorm <- qsmooth(object = assay(bodymapRat), 
                  groupFactor = colData(bodymapRat)$organ)
qsmoothPlotWeights(qsNorm)
}
