% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{qsmoothWeights}
\alias{qsmoothWeights}
\alias{qsmoothWeights,qsmooth-method}
\title{Generic function that returns the qsmooth weights}
\usage{
qsmoothWeights(object)

\S4method{qsmoothWeights}{qsmooth}(object)
}
\arguments{
\item{object}{an object of class \code{qsmooth}.}
}
\value{
The weights calculated for each feature 
after applying smoothed quantile normalization.
}
\description{
Given a qsmooth object, this function returns the 
qsmooth weights

Accessors for the 'qsmoothWeights' slot of a qsmooth object.
}
\examples{
library(SummarizedExperiment)
library(bodymapRat)
bm_dat <- bodymapRat()

# select lung and liver samples, stage 21 weeks, and bio reps
keep_columns = (colData(bm_dat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bm_dat)$stage == 21) & 
         (colData(bm_dat)$techRep == 1)
keep_rows = rowMeans(assay(bm_dat)) > 10 # Filter out low counts
bm_dat <- bm_dat[keep_rows,keep_columns]

qs_norm <- qsmooth(object = assay(bm_dat), 
                  group_factor = colData(bm_dat)$organ)
qsmoothWeights(qs_norm)

}
