% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{qsmoothWeights}
\alias{qsmoothWeights}
\alias{qsmoothWeights,qsmooth-method}
\alias{qsmoothWeights.qsmooth}
\title{Accessors for the 'qsmoothWeights' slot of a qsmooth object.}
\usage{
qsmoothWeights(object)

\S4method{qsmoothWeights}{qsmooth}(object)

\S4method{qsmoothWeights}{qsmooth}(object)
}
\arguments{
\item{object}{a \code{qsmooth} object}

\item{...}{other}
}
\description{
Accessors for the 'qsmoothWeights' slot of a qsmooth object.
}

