% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{qsmoothWeights}
\alias{qsmoothWeights}
\alias{qsmoothWeights,qsmooth-method}
\title{Generic function that returns the qsmooth weights}
\usage{
qsmoothWeights(object)

\S4method{qsmoothWeights}{qsmooth}(object)
}
\arguments{
\item{object}{an object of class \code{qsmooth}.}
}
\value{
The weights calculated for each feature 
after applying smoothed quantile normalization.
}
\description{
Given a qsmooth object, this function returns the 
qsmooth weights

Accessors for the 'qsmoothWeights' slot of a qsmooth object.
}
\examples{
library(SummarizedExperiment)
library(bodymapRat)
data(bodymapRat)

# select lung and liver samples, stage 21 weeks, and bio reps
keepColumns = (colData(bodymapRat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bodymapRat)$stage == 21) & 
         (colData(bodymapRat)$techRep == 1)
keepRows = rowMeans(assay(bodymapRat)) > 10 # Filter out low counts
bodymapRat <- bodymapRat[keepRows,keepColumns]

qsNorm <- qsmooth(object = assay(bodymapRat), 
                  groupFactor = colData(bodymapRat)$organ)
qsmoothWeights(qsNorm)

}
