% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qstats.r
\name{qstats}
\alias{qstats}
\title{Compute quantile statistics}
\usage{
qstats(exprs, groups, window)
}
\arguments{
\item{exprs}{for counts use log2(raw counts + 1)), for MA use log2(raw intensities)}

\item{groups}{groups to which samples belong (character vector)}

\item{window}{window size for running median as a fraction on the number of rows of exprs}
}
\description{
Compute quantile statistics
}

