% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstats.R
\docType{methods}
\name{qstats}
\alias{qstats}
\title{qstats}
\usage{
qstats(object, groupFactor, window = 0.05)
}
\arguments{
\item{object}{an object which is a data frame or 
matrix with observations (e.g. probes or genes) on 
the rows and samples as the columns.}

\item{groupFactor}{a group level continuous or categorial 
covariate associated with each sample or column in the 
\code{object}. The order of the \code{groupFactor} must 
match the order of the columns in \code{object}.}

\item{window}{window size for running median which is 
a fraction of the number of rows in \code{object}. 
Default is 0.05.}
}
\value{
A list of quantile statistics including 
\item{Q}{sample quantiles}
\item{Qref}{reference quantile}
\item{Qhat}{linear model fit at each quantile}
\item{SST}{total sum of squares}
\item{SSB}{between sum of squares}
\item{SSE}{within sum of squares}
\item{roughWeights}{SSE / SST}
\item{smoothWeights}{smoothed weights computed using a 
running median with a given \code{window} size.}
}
\description{
This function is a helper function that 
computes quantile statistics for the function 
\code{qsmooth}.
}
\examples{
library(SummarizedExperiment)
library(bodymapRat)
data(bodymapRat)

# select lung and liver samples, stage 21 weeks, and bio reps
keepColumns = (colData(bodymapRat)$organ \%in\% c("Lung", "Liver")) & 
         (colData(bodymapRat)$stage == 21) & 
         (colData(bodymapRat)$techRep == 1)
keepRows = rowMeans(assay(bodymapRat)) > 10 # Filter out low counts
bodymapRat <- bodymapRat[keepRows,keepColumns]

qs <- qstats(object = assay(bodymapRat), 
             groupFactor = colData(bodymapRat)$organ, 
             window = 0.05)

}
