% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulateCounts.r
\name{simulateCounts}
\alias{simulateCounts}
\title{Simulate RNA-seq counts via Tukey's G transformation and Negative Binomial distribution}
\usage{
simulateCounts(mu, groupSize, disp = 0.3, groupFC = NULL,
  techShapeBias = NULL, scaleBias = NULL)
}
\arguments{
\item{mu}{base counts}

\item{groupSize}{vector containing each group size. Names of groupSize should be the group names}

\item{disp}{dispersion parameter for NB distribution}

\item{groupFC}{the fc for every gene in every group (use first group as reference).}

\item{techShapeBias}{a vector of length sum(groupSize) indicating g shape parameters}

\item{scaleBais}{a vector of length sum(groupSize) indicating scaling factors}
}
\description{
Simulate RNA-seq counts via Tukey's G transformation and Negative Binomial distribution
}

